+++ selfdrive/controls/lib/longitudinal_planner.py
@@ -16,7 +16,7 @@ from openpilot.selfdrive.controls.lib.drive_helpers import CONTROL_N, get_speed_
 from openpilot.selfdrive.car.cruise import V_CRUISE_MAX, V_CRUISE_UNSET
 from openpilot.common.swaglog import cloudlog
 
-from sunnypilot.selfdrive.controls.dec.dec_longitudinal_planner import DecLongitudinalPlanner
+from openpilot.sunnypilot.selfdrive.controls.lib.dec_longitudinal_planner import DecLongitudinalPlanner
 
 LON_MPC_STEP = 0.2  # first step is 0.2s
 A_CRUISE_MIN = -1.2
